// src/lib/game-data/silk-data/player-data/player-data-silk.generated.ts
export type PlayerDataFieldValueOfTypeBoolSilk = boolean;
// ...
export type PlayerDataFieldValueOfTypeListOfStringSilk = string[];
export type PlayerDataFieldNameOfTypeListOfStringSilk = 'BelltownCouriersGenericQuests' | 'unlockedBossScenes';
export interface PlayerDataFieldOfTypeListOfStringSilk {
	id: number;
	name: PlayerDataFieldNameOfTypeListOfStringSilk;
	type: 'list<string>';
}

export type PlayerDataFieldValueOfTypeStringSilk = string;
export type PlayerDataFieldNameOfTypeStringSilk =
	| 'BelltownCouriersLastCompletedQuest'
	| 'BelltownHousePlayingInfo_RelicName'
	| 'CaravanSpiderTargetScene'
	| 'CrowSummonsAppearedScene'
	| 'CurrentCrestID'
	| 'HeroCorpseScene'
	| 'LastSetFieldName'
	| 'LibrarianPlayingInfo_RelicName'
	| 'MazeEntranceDoor'
	| 'MazeEntranceInitialDoor'
	| 'MazeEntranceInitialScene'
	| 'MazeEntranceScene'
	| 'PreviousCrestID'
	| 'PreviousMazeDoor'
	| 'PreviousMazeScene'
	| 'PreviousMazeTargetDoor'
	| 'bossReturnEntryGate'
	| 'currentArea'
	| 'currentBossStatueCompletionKey'
	| 'date'
	| 'dreamReturnScene'
	| 'nextScene'
	| 'nonLethalRespawnMarker'
	| 'nonLethalRespawnScene'
	| 'respawnMarkerName'
	| 'respawnScene'
	| 'tempRespawnMarker'
	| 'tempRespawnScene'
	| 'version';
export interface PlayerDataFieldOfTypeStringSilk {
	id: number;
	name: PlayerDataFieldNameOfTypeStringSilk;
	type: 'string';
}

export type PlayerDataFieldValueOfTypeIntSilk = number;
export type PlayerDataFieldNameOfTypeIntSilk =
	| 'BelltownCouriersPurchasedDlgBitmask'
	| 'BelltownDoctorConvo'
	| 'BelltownGreeterConvo'
	| 'BelltownHermitEnslavedConvo'
	| 'BelltownHermitSavedConvo'
	| 'BlueAssistantBloodCount'
	| 'BonebottomBellwayPilgrimState'
	| 'CorrectMazeDoorsEntered'
	| 'ForgeDaughterTalkState'
	| 'GrowstoneState'
	| 'GrubFarmerSilkGrubsSold'
	| 'HalfwayDrinksPurchased'
	| 'HeroCorpseMoneyPool'
	| 'IncorrectMazeDoorsEntered'
	| 'MaggotCharmHits'
	| 'PilgrimsRestShopIdleTalkState'
	| 'PinGalleryWallet'
	| 'PreMemoryState_Health'
	| 'PreMemoryState_Rosaries'
	| 'PreMemoryState_ShellShards'
	| 'PreMemoryState_Silk'
	| 'QuillState'
	| 'RevisionBreak'
	| 'SeamstressIdleTalkState'
	| 'ShellShards'
	| 'SprintMasterCurrentRace'
	| 'TempGeoStore'
	| 'TempShellShardStore'
	| 'ToolKitUpgrades'
	| 'ToolPouchUpgrades'
	| 'attunement'
	| 'attunementLevel'
	| 'beamDamage'
	| 'belltownCrowd'
	| 'bonetownCrowd'
	| 'bonetownPilgrimHornedCount'
	| 'bonetownPilgrimRoundCount'
	| 'bossStatueTargetLevel'
	| 'cloakOdour_slabFly'
	| 'currentInvPane'
	| 'dicePilgrimBank'
	| 'dicePilgrimState'
	| 'druidMossBerriesSold'
	| 'enclaveLevel'
	| 'enemyGroupAnt04'
	| 'fleaGames_bouncing_highscore'
	| 'fleaGames_dodging_highscore'
	| 'fleaGames_juggling_highscore'
	| 'garmondEncounters_act3'
	| 'geo'
	| 'gillyLocation'
	| 'gillyLocationAct3'
	| 'grindleChestLocation'
	| 'grindleSlabSequence'
	| 'grubFarmLevel'
	| 'halfwayCrowEnemyGroup'
	| 'halfwayCrowd'
	| 'health'
	| 'healthBlue'
	| 'heartPieces'
	| 'heroState_attackCount'
	| 'heroState_downspikeInvulnerabilitySteps'
	| 'heroState_invulnerableCount'
	| 'heroState_throwingToolVertical'
	| 'heroState_toolThrowCount'
	| 'mapKeyPref'
	| 'mapperFightGroup'
	| 'mapperLocationAct3'
	| 'maxHealth'
	| 'maxHealthBase'
	| 'nailRange'
	| 'nailUpgrades'
	| 'nonLethalRespawnType'
	| 'pilbyFriendship'
	| 'pilgrimGroupBonegrave'
	| 'pilgrimGroupGreymoorField'
	| 'pilgrimGroupShellgrave'
	| 'pilgrimRestCrowd'
	| 'pinGalleriesCompleted'
	| 'prevHealth'
	| 'previousDarkness'
	| 'profileID'
	| 'respawnType'
	| 'sethConvo'
	| 'shermaPos'
	| 'shermaWoundedPilgrim'
	| 'silk'
	| 'silkMax'
	| 'silkParts'
	| 'silkRegenMax'
	| 'silkSpecialLevel'
	| 'silkSpoolParts'
	| 'songChevalierEncounters'
	| 'spinnerEncounter'
	| 'splinterQueenSproutTimer'
	| 'tempRespawnType'
	| 'vampireGnatDeaths'
	| 'whiteCloverPos';
export interface PlayerDataFieldOfTypeIntSilk {
	id: number;
	name: PlayerDataFieldNameOfTypeIntSilk;
	type: 'int';
}

export type PlayerDataFieldValueOfTypeEnumSilk = number;
export type PlayerDataFieldNameOfTypeEnumSilk =
	| 'BelltownHouseColour'
	| 'BelltownHouseState'
	| 'CaravanTroupeLocation'
	| 'CompletedEndings'
	| 'EnclaveStateNPCShortHorned'
	| 'EnclaveStateNPCStandard'
	| 'EnclaveStateNPCTall'
	| 'EnclaveStatePilgrimSmall'
	| 'EnclaveState_songKnightFan'
	| 'FastTravelNPCLocation'
	| 'GreenPrinceLocation'
	| 'HeroCorpseType'
	| 'LastCompletedEnding'
	| 'SethNpcLocation'
	| 'environmentType'
	| 'extraRestZone'
	| 'hazardRespawnFacing'
	| 'mapZone'
	| 'permadeathMode'
	| 'gameManager_GameState';
export interface PlayerDataFieldOfTypeEnumSilk {
	id: number;
	name: PlayerDataFieldNameOfTypeEnumSilk;
	type: 'enum';
}

export type PlayerDataFieldValueOfTypeFloatSilk = number;
export type PlayerDataFieldNameOfTypeFloatSilk =
	| 'BelltownHousePlayingInfo_StartTime'
	| 'CaravanSpiderTravelDirection'
	| 'CrawbellTimer'
	| 'FisherWalkerIdleTimeLeft'
	| 'FisherWalkerTimer'
	| 'GrowstoneTimer'
	| 'LibrarianPlayingInfo_StartTime'
	| 'completionPercentage'
	| 'greyWarriorDeathX'
	| 'grubFarmerTimer'
	| 'laceCorpsePosX'
	| 'laceCorpseScaleX'
	| 'playTime';
export interface PlayerDataFieldOfTypeFloatSilk {
	id: number;
	name: PlayerDataFieldNameOfTypeFloatSilk;
	type: 'float';
}

export type PlayerDataFieldValueOfTypeArrayOfIntSilk = number[];
export type PlayerDataFieldNameOfTypeArrayOfIntSilk =
	| 'CrawbellCurrency'
	| 'CrawbellCurrencyCaps'
	| 'GrubFarmerMimicValueList'
	| 'mossBerryValueList';
export interface PlayerDataFieldOfTypeArrayOfIntSilk {
	id: number;
	name: PlayerDataFieldNameOfTypeArrayOfIntSilk;
	type: 'int[]';
}

export type PlayerDataFieldValueOfTypeListOfIntSilk = number[];
export type PlayerDataFieldNameOfTypeListOfIntSilk = 'FleasCollectedTargetOrder';
export interface PlayerDataFieldOfTypeListOfIntSilk {
	id: number;
	name: PlayerDataFieldNameOfTypeListOfIntSilk;
	type: 'list<int>';
}

export type PlayerDataFieldValueOfTypeGuidSilk = string;
export type PlayerDataFieldNameOfTypeGuidSilk = 'HeroCorpseMarkerGuid';
export interface PlayerDataFieldOfTypeGuidSilk {
	id: number;
	name: PlayerDataFieldNameOfTypeGuidSilk;
	type: 'guid';
}

export type PlayerDataFieldValueOfTypeVector2Silk = Vector2;
export type PlayerDataFieldNameOfTypeVector2Silk = 'HeroDeathScenePos' | 'HeroDeathSceneSize';
export interface PlayerDataFieldOfTypeVector2Silk {
	id: number;
	name: PlayerDataFieldNameOfTypeVector2Silk;
	type: 'vector2';
}

export type PlayerDataFieldValueOfTypeDictionaryOfString_boolSilk = Map<string, boolean>;
export type PlayerDataFieldNameOfTypeDictionaryOfString_boolSilk = 'SteelQuestSpots';
export interface PlayerDataFieldOfTypeDictionaryOfString_boolSilk {
	id: number;
	name: PlayerDataFieldNameOfTypeDictionaryOfString_boolSilk;
	type: 'dictionary<string,bool>';
}

export type PlayerDataFieldValueOfTypeVector3Silk = Vector3;
export type PlayerDataFieldNameOfTypeVector3Silk = 'hazardRespawnLocation';
export interface PlayerDataFieldOfTypeVector3Silk {
	id: number;
	name: PlayerDataFieldNameOfTypeVector3Silk;
	type: 'vector3';
}

export type PlayerDataFieldValueOfTypeUlongSilk = number;
export type PlayerDataFieldNameOfTypeUlongSilk =
	| 'memoryOrbs_Clover_18_A'
	| 'memoryOrbs_Clover_18_B'
	| 'memoryOrbs_Clover_18_C'
	| 'memoryOrbs_Clover_18_D'
	| 'memoryOrbs_Clover_18_E'
	| 'memoryOrbs_Clover_19';
export interface PlayerDataFieldOfTypeUlongSilk {
	id: number;
	name: PlayerDataFieldNameOfTypeUlongSilk;
	type: 'ulong';
}

export type PlayerDataFieldValueOfTypeArrayOfWrappedvector2listSilk = Vector2[][];
export type PlayerDataFieldNameOfTypeArrayOfWrappedvector2listSilk = 'placedMarkers';
export interface PlayerDataFieldOfTypeArrayOfWrappedvector2listSilk {
	id: number;
	name: PlayerDataFieldNameOfTypeArrayOfWrappedvector2listSilk;
	type: 'wrappedvector2list[]';
}

export type PlayerDataFieldValueOfTypeHashsetOfStringSilk = Set<string>;
export type PlayerDataFieldNameOfTypeHashsetOfStringSilk =
	| 'scenesEncounteredBench'
	| 'scenesEncounteredCocoon'
	| 'scenesMapped'
	| 'scenesVisited';
export interface PlayerDataFieldOfTypeHashsetOfStringSilk {
	id: number;
	name: PlayerDataFieldNameOfTypeHashsetOfStringSilk;
	type: 'hashset<string>';
}

export type PlayerDataFieldValueOfTypeDictionaryOfString_intSilk = Map<string, number>;
export type PlayerDataFieldNameOfTypeDictionaryOfString_intSilk = 'toolAmountsOverride';
export interface PlayerDataFieldOfTypeDictionaryOfString_intSilk {
	id: number;
	name: PlayerDataFieldNameOfTypeDictionaryOfString_intSilk;
	type: 'dictionary<string,int>';
}

export type PlayerDataFieldValueOfTypeCollectableItemsDataSilk = Map<string, CollectableItemsDataSilk>;
export type PlayerDataFieldNameOfTypeCollectableItemsDataSilk = 'Collectables';
export interface PlayerDataFieldOfTypeCollectableItemsDataSilk {
	id: number;
	name: PlayerDataFieldNameOfTypeCollectableItemsDataSilk;
	type: 'CollectableItemsData';
}

export type PlayerDataFieldValueOfTypeEnemyJournalKillDataSilk = Map<string, EnemyJournalKillDataSilk>;
export type PlayerDataFieldNameOfTypeEnemyJournalKillDataSilk = 'EnemyJournalKillData';
export interface PlayerDataFieldOfTypeEnemyJournalKillDataSilk {
	id: number;
	name: PlayerDataFieldNameOfTypeEnemyJournalKillDataSilk;
	type: 'EnemyJournalKillData';
}

export type PlayerDataFieldValueOfTypeMateriumItemsDataSilk = Map<string, MateriumItemsDataSilk>;
export type PlayerDataFieldNameOfTypeMateriumItemsDataSilk = 'MateriumCollected';
export interface PlayerDataFieldOfTypeMateriumItemsDataSilk {
	id: number;
	name: PlayerDataFieldNameOfTypeMateriumItemsDataSilk;
	type: 'MateriumItemsData';
}

export type PlayerDataFieldValueOfTypeCollectableMementosDataSilk = Map<string, CollectableMementosDataSilk>;
export type PlayerDataFieldNameOfTypeCollectableMementosDataSilk = 'MementosDeposited';
export interface PlayerDataFieldOfTypeCollectableMementosDataSilk {
	id: number;
	name: PlayerDataFieldNameOfTypeCollectableMementosDataSilk;
	type: 'CollectableMementosData';
}

export type PlayerDataFieldValueOfTypeQuestCompletionDataSilk = Map<string, QuestCompletionDataSilk>;
export type PlayerDataFieldNameOfTypeQuestCompletionDataSilk = 'QuestCompletionData';
export interface PlayerDataFieldOfTypeQuestCompletionDataSilk {
	id: number;
	name: PlayerDataFieldNameOfTypeQuestCompletionDataSilk;
	type: 'QuestCompletionData';
}

export type PlayerDataFieldValueOfTypeQuestRumourDataSilk = Map<string, QuestRumourDataSilk>;
export type PlayerDataFieldNameOfTypeQuestRumourDataSilk = 'QuestRumourData';
export interface PlayerDataFieldOfTypeQuestRumourDataSilk {
	id: number;
	name: PlayerDataFieldNameOfTypeQuestRumourDataSilk;
	type: 'QuestRumourData';
}

export type PlayerDataFieldValueOfTypeCollectableRelicsDataSilk = Map<string, CollectableRelicsDataSilk>;
export type PlayerDataFieldNameOfTypeCollectableRelicsDataSilk = 'Relics';
export interface PlayerDataFieldOfTypeCollectableRelicsDataSilk {
	id: number;
	name: PlayerDataFieldNameOfTypeCollectableRelicsDataSilk;
	type: 'CollectableRelicsData';
}

export type PlayerDataFieldValueOfTypeToolCrestsDataSilk = Map<string, ToolCrestsDataSilk>;
export type PlayerDataFieldNameOfTypeToolCrestsDataSilk = 'ToolEquips';
export interface PlayerDataFieldOfTypeToolCrestsDataSilk {
	id: number;
	name: PlayerDataFieldNameOfTypeToolCrestsDataSilk;
	type: 'ToolCrestsData';
}

export type PlayerDataFieldValueOfTypeToolItemLiquidsDataSilk = Map<string, ToolItemLiquidsDataSilk>;
export type PlayerDataFieldNameOfTypeToolItemLiquidsDataSilk = 'ToolLiquids';
export interface PlayerDataFieldOfTypeToolItemLiquidsDataSilk {
	id: number;
	name: PlayerDataFieldNameOfTypeToolItemLiquidsDataSilk;
	type: 'ToolItemLiquidsData';
}

export type PlayerDataFieldValueOfTypeToolItemsDataSilk = Map<string, ToolItemsDataSilk>;
export type PlayerDataFieldNameOfTypeToolItemsDataSilk = 'Tools';
export interface PlayerDataFieldOfTypeToolItemsDataSilk {
	id: number;
	name: PlayerDataFieldNameOfTypeToolItemsDataSilk;
	type: 'ToolItemsData';
}

export type PlayerDataFieldValueOfTypeFloatingCrestSlotsDataSilk = unknown;
export type PlayerDataFieldNameOfTypeFloatingCrestSlotsDataSilk = 'ExtraToolEquips';
export interface PlayerDataFieldOfTypeFloatingCrestSlotsDataSilk {
	id: number;
	name: PlayerDataFieldNameOfTypeFloatingCrestSlotsDataSilk;
	type: 'FloatingCrestSlotsData';
}

export type PlayerDataFieldValueOfTypeListOfPlayerstory_eventinfoSilk = unknown;
export type PlayerDataFieldNameOfTypeListOfPlayerstory_eventinfoSilk = 'StoryEvents';
export interface PlayerDataFieldOfTypeListOfPlayerstory_eventinfoSilk {
	id: number;
	name: PlayerDataFieldNameOfTypeListOfPlayerstory_eventinfoSilk;
	type: 'list<playerstory.eventinfo>';
}

export type PlayerDataFieldNameSilk =
	| PlayerDataFieldNameOfTypeBoolSilk
	| PlayerDataFieldNameOfTypeListOfStringSilk
	| PlayerDataFieldNameOfTypeStringSilk
	| PlayerDataFieldNameOfTypeIntSilk
	| PlayerDataFieldNameOfTypeEnumSilk
	| PlayerDataFieldNameOfTypeFloatSilk
	| PlayerDataFieldNameOfTypeArrayOfIntSilk
	| PlayerDataFieldNameOfTypeListOfIntSilk
	| PlayerDataFieldNameOfTypeGuidSilk
	| PlayerDataFieldNameOfTypeVector2Silk
	| PlayerDataFieldNameOfTypeDictionaryOfString_boolSilk
	| PlayerDataFieldNameOfTypeVector3Silk
	| PlayerDataFieldNameOfTypeUlongSilk
	| PlayerDataFieldNameOfTypeArrayOfWrappedvector2listSilk
	| PlayerDataFieldNameOfTypeHashsetOfStringSilk
	| PlayerDataFieldNameOfTypeDictionaryOfString_intSilk
	| PlayerDataFieldNameOfTypeCollectableItemsDataSilk
	| PlayerDataFieldNameOfTypeEnemyJournalKillDataSilk
	| PlayerDataFieldNameOfTypeMateriumItemsDataSilk
	| PlayerDataFieldNameOfTypeCollectableMementosDataSilk
	| PlayerDataFieldNameOfTypeQuestCompletionDataSilk
	| PlayerDataFieldNameOfTypeQuestRumourDataSilk
	| PlayerDataFieldNameOfTypeCollectableRelicsDataSilk
	| PlayerDataFieldNameOfTypeToolCrestsDataSilk
	| PlayerDataFieldNameOfTypeToolItemLiquidsDataSilk
	| PlayerDataFieldNameOfTypeToolItemsDataSilk
	| PlayerDataFieldNameOfTypeFloatingCrestSlotsDataSilk
	| PlayerDataFieldNameOfTypeListOfPlayerstory_eventinfoSilk;
// ...
export type PlayerDataFieldValueSilk<TField extends PlayerDataFieldNameSilk> =
	TField extends PlayerDataFieldNameOfTypeBoolSilk
		? PlayerDataFieldValueOfTypeBoolSilk
		: TField extends PlayerDataFieldNameOfTypeListOfStringSilk
			? PlayerDataFieldValueOfTypeListOfStringSilk
			: TField extends PlayerDataFieldNameOfTypeStringSilk
				? PlayerDataFieldValueOfTypeStringSilk
				: TField extends PlayerDataFieldNameOfTypeIntSilk
					? PlayerDataFieldValueOfTypeIntSilk
					: TField extends PlayerDataFieldNameOfTypeEnumSilk
						? PlayerDataFieldValueOfTypeEnumSilk
						: TField extends PlayerDataFieldNameOfTypeFloatSilk
							? PlayerDataFieldValueOfTypeFloatSilk
							: TField extends PlayerDataFieldNameOfTypeArrayOfIntSilk
								? PlayerDataFieldValueOfTypeArrayOfIntSilk
								: TField extends PlayerDataFieldNameOfTypeListOfIntSilk
									? PlayerDataFieldValueOfTypeListOfIntSilk
									: TField extends PlayerDataFieldNameOfTypeGuidSilk
										? PlayerDataFieldValueOfTypeGuidSilk
										: TField extends PlayerDataFieldNameOfTypeVector2Silk
											? PlayerDataFieldValueOfTypeVector2Silk
											: TField extends PlayerDataFieldNameOfTypeDictionaryOfString_boolSilk
												? PlayerDataFieldValueOfTypeDictionaryOfString_boolSilk
												: TField extends PlayerDataFieldNameOfTypeVector3Silk
													? PlayerDataFieldValueOfTypeVector3Silk
													: TField extends PlayerDataFieldNameOfTypeUlongSilk
														? PlayerDataFieldValueOfTypeUlongSilk
														: TField extends PlayerDataFieldNameOfTypeArrayOfWrappedvector2listSilk
															? PlayerDataFieldValueOfTypeArrayOfWrappedvector2listSilk
															: TField extends PlayerDataFieldNameOfTypeHashsetOfStringSilk
																? PlayerDataFieldValueOfTypeHashsetOfStringSilk
																: TField extends PlayerDataFieldNameOfTypeDictionaryOfString_intSilk
																	? PlayerDataFieldValueOfTypeDictionaryOfString_intSilk
																	: TField extends PlayerDataFieldNameOfTypeCollectableItemsDataSilk
																		? PlayerDataFieldValueOfTypeCollectableItemsDataSilk
																		: TField extends PlayerDataFieldNameOfTypeEnemyJournalKillDataSilk
																			? PlayerDataFieldValueOfTypeEnemyJournalKillDataSilk
																			: TField extends PlayerDataFieldNameOfTypeMateriumItemsDataSilk
																				? PlayerDataFieldValueOfTypeMateriumItemsDataSilk
																				: TField extends PlayerDataFieldNameOfTypeCollectableMementosDataSilk
																					? PlayerDataFieldValueOfTypeCollectableMementosDataSilk
																					: TField extends PlayerDataFieldNameOfTypeQuestCompletionDataSilk
																						? PlayerDataFieldValueOfTypeQuestCompletionDataSilk
																						: TField extends PlayerDataFieldNameOfTypeQuestRumourDataSilk
																							? PlayerDataFieldValueOfTypeQuestRumourDataSilk
																							: TField extends PlayerDataFieldNameOfTypeCollectableRelicsDataSilk
																								? PlayerDataFieldValueOfTypeCollectableRelicsDataSilk
																								: TField extends PlayerDataFieldNameOfTypeToolCrestsDataSilk
																									? PlayerDataFieldValueOfTypeToolCrestsDataSilk
																									: TField extends PlayerDataFieldNameOfTypeToolItemLiquidsDataSilk
																										? PlayerDataFieldValueOfTypeToolItemLiquidsDataSilk
																										: TField extends PlayerDataFieldNameOfTypeToolItemsDataSilk
																											? PlayerDataFieldValueOfTypeToolItemsDataSilk
																											: TField extends PlayerDataFieldNameOfTypeFloatingCrestSlotsDataSilk
																												? PlayerDataFieldValueOfTypeFloatingCrestSlotsDataSilk
																												: TField extends PlayerDataFieldNameOfTypeListOfPlayerstory_eventinfoSilk
																													? PlayerDataFieldValueOfTypeListOfPlayerstory_eventinfoSilk
																													: never;